+++ CHANGELOG.md
@@ -4,13 +4,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- VIMEET-2185 - Add waiting image/video to webinar
+
 ## [2.71.3] - 2020-11-25
 ### Fixed
 - MV-282 - Delete header on player video SQY78
 
 ## [2.71.2] - 2020-11-23
 ### Fixed
-- MV-278 - Video player SQY78 
+- MV-278 - Video player SQY78
 
 ## [2.71.1] - 2020-11-22
 ### Fixed